move setup_mikrotik_entry to __init__.py
